+++ src/corebluetooth/peripheral.rs
@@ -48,6 +48,12 @@ use uuid::Uuid;
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct PeripheralId(Uuid);
 
+impl Display for PeripheralId {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        Display::fmt(&self.0, f)
+    }
+}
+
 /// Implementation of [api::Peripheral](crate::api::Peripheral).
 #[derive(Clone)]
 pub struct Peripheral {
@@ -151,7 +157,7 @@ impl Peripheral {
                     }
                     Some(CBPeripheralEvent::Disconnected) => (),
                     None => {
-                        error!("Event receiver died, breaking out of corebluetooth device loop.");
+                        info!("Event receiver died, breaking out of corebluetooth device loop.");
                         break;
                     }
                 }
@@ -250,7 +256,23 @@ impl api::Peripheral for Peripheral {
     }
 
     async fn disconnect(&self) -> Result<()> {
-        // TODO
+        let fut = CoreBluetoothReplyFuture::default();
+        self.shared
+            .message_sender
+            .to_owned()
+            .send(CoreBluetoothMessage::DisconnectDevice {
+                peripheral_uuid: self.shared.uuid,
+                future: fut.get_state_clone(),
+            })
+            .await?;
+        match fut.await {
+            CoreBluetoothReply::Ok => {
+                self.shared
+                    .emit_event(CentralEvent::DeviceDisconnected(self.shared.uuid.into()));
+                trace!("Device disconnected!");
+            }
+            _ => error!("Shouldn't get anything but Ok!"),
+        }
         Ok(())
     }