+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+# 0.10.5 (2023-04-13)
+
+## Features
+
+- Add RSSI readings for Android
+
+## Bugfixes
+
+- Link conditionally against macOS AppKit based on platform
+- Improve error propagation on Windows
+- Reset connected state to false on disconnect on windows
+- Set DuplicateData = true for bluez
+
+# 0.10.4 (2022-11-27)
+
+## Bugfixes
+
+- Change common CoreBluetooth log message from error to info level
+
+# 0.10.3 (2022-11-05)
+
+## Bugfixes
+
+- Add PeripheralId Display implementation for Android PeripheralId
+
+# 0.10.2 (2022-10-30)
+
+## Features
+
+- Implement Display on PeripheralId
+
+## Bugfixes
+
+- Fix issues with panics on device disconnect on macOS
+
+# 0.10.1 (2022-09-23)
+
+## Features
+
+- Add ability to disconnect devices on macOS/iOS
+
 # 0.10.0 (2022-07-30)
 
 ## Features